Better workflow now!

+++ pmos_sideload.sh
@@ -1,4 +1,13 @@
 #!/usr/bin/sh
+
+set -Eeuo pipefail
+
+if [ $# -ne 1 ]
+then
+	echo "usage: $0 ssh-device-name"
+	exit 1
+fi
+
 # clean dir
 rm -rf ~/.local/var/pmbootstrap/cache_git/pmaports/testing/
 # stay updated!
@@ -6,19 +15,22 @@ echo agent setup...
 eval $(ssh-agent)
 ssh-add
 
-echo pulling...
+echo pmbootstrap pulling...
 pmbootstrap pull
 git push
 
 mkdir -p ~/.local/var/pmbootstrap/cache_git/pmaports/testing/dewduct
-cp pmaport/APKBUILD ~/.local/var/pmbootstrap/cache_git/pmaports/testing/dewduct
+cp aport/APKBUILD_dev ~/.local/var/pmbootstrap/cache_git/pmaports/testing/dewduct/APKBUILD
 
 # prepare the apk
 echo checksum...
 pmbootstrap checksum dewduct
-cp ~/.local/var/pmbootstrap/cache_git/pmaports/testing/dewduct/APKBUILD pmaport/
+cp ~/.local/var/pmbootstrap/cache_git/pmaports/testing/dewduct/APKBUILD aport/APKBUILD_dev
 echo build...
-pmbootstrap build --arch aarch64 dewduct --force
+# hack: to avoid rsync-ing target folder (tens of gigs!) into the chrootsx
+mv target .git/
+pmbootstrap build --arch aarch64 dewduct --src="$PWD"
+mv .git/target target
 # push it
 echo sideload...
 scp ~/.local/var/pmbootstrap/packages/edge/aarch64/dewduct*.apk \
